+++ packages/babel-plugin/__tests__/fixtures/inline-macro/called-multiple/output.js
@@ -1,4 +1,4 @@
-import { create } from "@stylexjs/stylex";
+import { create, props as _props } from "@stylexjs/stylex";
 import stylex from "@stylexjs/stylex";
 const styles = create({
   base: {